+++ shared/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.24] - 2023-02-09
+
+ - Add `ConsensusStateSync` interface that is implemented by the consensus module
+
 ## [0.0.0.23] - 2023-02-07
 
 - Added GITHUB_WIKI tags where it was missing